Moved the bar function into utils. Added comments and docstrings. Fixed the iter_uuids method
